+++ src/Utils/stringUtils/extension.d.ts
@@ -1,0 +1,11 @@
+declare global {
+    interface String {
+        replaceAllChars(str: Array<string>, replaceChar: string): string;
+        firstCharToUpper(): string;
+        firstCharToLower(): string;
+        transBigCamel(splitChars: string[]): string;
+        transSmallCamel(splitChars: string[]): string;
+    }
+}
+
+export {};